// app0/src/app/interfaces/response/cat.ts
import {PropFullInterface} from './prop';

export interface CatInterface {
    catID: number;
    name: string;
    slug: string;
    parentID: number;
    pos: number;
    isDisabled: boolean;
    priceAlias: string;
    priceSuffix: string;
    titleHelp: string;
    titleComment: string;
    isAutogenerateTitle: boolean;
}

export interface CatTreeInterface extends CatInterface, ChildesInterface {
}

export interface CatFullInterface extends CatInterface, PropsFullInterface {
}

export interface CatTreeFullInterface extends CatFullInterface, ChildesFullInterface {
}

export interface CatWithDeepInterface extends CatInterface, WithDeepInterface {
}

interface ChildesInterface {
    childes: CatTreeInterface[];
}

interface ChildesFullInterface {
    childes: CatTreeFullInterface[];
}

interface PropsFullInterface {
    props: PropFullInterface[];
}

interface WithDeepInterface {
    deep: number;
}